Improve test tool methods handling
Move boilerplate wrapping code from the sub exported by `Test::Async`
into a tool method wrapper.

The wrapper is installed as late as possible, when all methods and
candidates were installed and incorporated.

Ultimately, it is now as safe to invoke `test-suite.test-tool(...)` as
it was with just `test-tool(...)` in a test suite file. In other words:

    ok ?$got, "test";
    test-suite.ok: $got, "test";

are now fully equivalent.