Make the module Rakudo 2019.11 compatible
Earlier versions support is not planned.

The key change to make this compatibility possible is switching from
class-based test bundles to role-based. Unfortuantely, `EXPORTHOW`
doesn't work for roles and even if it would for later compiler versions,
it doesn't for the old ones. Thus, I had to pick up the trick from
`OO::Plugins` and define own slang.

The next thing changed is the inheritance model. I now create custom
`Test::Async::Metamodel::BundleClassHOW`-based class for each bundle
which is totally empty and only does bundle's role. I.e. it is a kind of
puning. The difference is that the custom class gets the last created
bundle class or the hub class as it's parent. This fixes a problem with
non-chaining dispatchers in old Rakudo compilers resulting in
impossibility for one multi-method to redispatch to a method of a next
parent breaking overriding of parent's protos.

Removed few recent syntax additions.