Make module play well with independent threads
Previously if a thread was created not by the job manager and a test
tool is invoked in the thread there was chances that it would compete
for the test tool stack. From now on if we detect that a call takes
place in a thread different of what is recorded on the test suite
object (its primary thread) and there is no `$*TEST-SUITE` variable
then the call is explicitly gets wrapped into a new job. It makes
calling tools in independent threads slower but safe.